+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * introduced scalar type hints and strict type checking
   * `stubbles\reflect\annotation\Annotations::all()` now returns a `\Generator` instead of an `array`
   * parameter annotations with empty parameter name like `@Foo{}` are now invalid and throw a `\ReflectionException`
+  * empty refined annotation types like `@Foo[]` are now invalid and throw a `\ReflectionException`
 
 
 7.0.0 (2016-01-11)